+++ system/platform-core/src/main/java/org/platformlambda/core/exception/AppException.java
@@ -29,6 +29,11 @@ public class AppException extends Exception {
         this.status = status;
     }
 
+    public AppException(int status, String message, Throwable cause) {
+        super(message, cause);
+        this.status = status;
+    }
+
     public int getStatus() {
         return status;
     }