+++ CHANGELOG.md
@@ -6,11 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
-## Version 2.2.2, 11/10/2021
+## Version 2.2.2, 11/12/2021
 
 ### Added
 
-User defined API authentication functions can be selected using custom HTTP request header
+1. User defined API authentication functions can be selected using custom HTTP request header
+2. Added "exception object" transport in EventEnvelope
 
 ### Removed